+++ CHANGELOG.md
@@ -4,8 +4,9 @@
 - Allow bots to send a 'typing' message
 - Messages will be sent via the Real-Team API if possible (not all message parameters are acceptable there)
 - Subsequent bot callbacks won't fire if an earlier one returns `false`
-- `SlackBotServer::Bot` now exposes `user`, `user_id`, `team`, and `team_id` methods
+- `SlackBotServer::Bot` now exposes `bot_user_name`, `bot_user_id`, `team_name`, and `team_id` methods
 - The logger can now be set via `SlackBotServer.logger=`
+- Access the underlying Slack client via the `SlackBotServer::Bot#client` method
 
 ### Changes
 - Swapped internal API library from `slack-api` to `slack-ruby-client`